+++ lib/processor/decode_test.go
@@ -1,0 +1,165 @@
+// Copyright (c) 2018 Ashley Jeffs
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package processor
+
+import (
+	"bytes"
+	"encoding/base64"
+	"os"
+	"reflect"
+	"testing"
+
+	"github.com/Jeffail/benthos/lib/log"
+	"github.com/Jeffail/benthos/lib/metrics"
+	"github.com/Jeffail/benthos/lib/types"
+)
+
+func TestDecodeBadAlgo(t *testing.T) {
+	conf := NewConfig()
+	conf.Decode.Scheme = "does not exist"
+
+	testLog := log.New(os.Stdout, log.Config{LogLevel: "NONE"})
+
+	_, err := NewDecode(conf, nil, testLog, metrics.DudType{})
+	if err == nil {
+		t.Error("Expected error from bad algo")
+	}
+}
+
+func TestDecodeBase64(t *testing.T) {
+	conf := NewConfig()
+	conf.Decode.Scheme = "base64"
+
+	testLog := log.New(os.Stdout, log.Config{LogLevel: "NONE"})
+
+	exp := [][]byte{
+		[]byte("hello world first part"),
+		[]byte("hello world second part"),
+		[]byte("third part"),
+		[]byte("fourth"),
+		[]byte("5"),
+	}
+
+	input := [][]byte{}
+
+	for i := range exp {
+		var buf bytes.Buffer
+
+		zw := base64.NewEncoder(base64.StdEncoding, &buf)
+		zw.Write(exp[i])
+		zw.Close()
+
+		input = append(input, buf.Bytes())
+	}
+
+	if reflect.DeepEqual(input, exp) {
+		t.Fatal("Input and exp output are the same")
+	}
+
+	proc, err := NewDecode(conf, nil, testLog, metrics.DudType{})
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	msgs, res := proc.ProcessMessage(types.NewMessage(input))
+	if len(msgs) != 1 {
+		t.Error("Decode failed")
+	} else if res != nil {
+		t.Errorf("Expected nil response: %v", res)
+	}
+	if act := msgs[0].GetAll(); !reflect.DeepEqual(exp, act) {
+		t.Errorf("Unexpected output: %s != %s", act, exp)
+	}
+}
+
+func TestDecodeIndexBounds(t *testing.T) {
+	conf := NewConfig()
+
+	testLog := log.New(os.Stdout, log.Config{LogLevel: "NONE"})
+
+	input := [][]byte{}
+
+	decoded := [][]byte{
+		[]byte("0"),
+		[]byte("1"),
+		[]byte("2"),
+		[]byte("3"),
+		[]byte("4"),
+	}
+
+	for i := range decoded {
+		var buf bytes.Buffer
+
+		zw := base64.NewEncoder(base64.StdEncoding, &buf)
+		zw.Write(decoded[i])
+		zw.Close()
+
+		input = append(input, buf.Bytes())
+	}
+
+	tests := map[int]int{
+		-5: 0,
+		-4: 1,
+		-3: 2,
+		-2: 3,
+		-1: 4,
+		0:  0,
+		1:  1,
+		2:  2,
+		3:  3,
+		4:  4,
+	}
+
+	for i, expIndex := range tests {
+		conf.Decode.Parts = []int{i}
+		proc, err := NewDecode(conf, nil, testLog, metrics.DudType{})
+		if err != nil {
+			t.Fatal(err)
+		}
+
+		msgs, res := proc.ProcessMessage(types.NewMessage(input))
+		if len(msgs) != 1 {
+			t.Errorf("Decode failed on index: %v", i)
+		} else if res != nil {
+			t.Errorf("Expected nil response: %v", res)
+		}
+		if exp, act := string(decoded[expIndex]), string(msgs[0].GetAll()[expIndex]); exp != act {
+			t.Errorf("Unexpected output for index %v: %v != %v", i, act, exp)
+		}
+	}
+}
+
+func TestDecodeEmpty(t *testing.T) {
+	conf := NewConfig()
+	conf.Decode.Parts = []int{0, 1}
+
+	testLog := log.New(os.Stdout, log.Config{LogLevel: "NONE"})
+	proc, err := NewDecode(conf, nil, testLog, metrics.DudType{})
+	if err != nil {
+		t.Error(err)
+		return
+	}
+
+	msgs, _ := proc.ProcessMessage(types.NewMessage([][]byte{}))
+	if len(msgs) > 0 {
+		t.Error("Expected failure with zero part message")
+	}
+}